refactor: Rename notify modules
